# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация API",
        "title": "Hack",
        "contact": {},
        "version": "1.0"
    },
    "host": "37.18.110.184:3000",
    "basePath": "/",
    "paths": {
...
        "/getUserInfo": {
            "get": {
                "description": "На вход получаю токен в хедере Authorization",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Проверка токена",
                "operationId": "UserIdentification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        },
...
        "/user/businessRegister": {
            "post": {
                "description": "Регистрация пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация арендодателя",
                "operationId": "landlordRegister",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "regDate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/golang-pkg_internal_auth.BusinessUserForRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/golang-pkg_internal_auth.BusinessUserForRegister"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "Авторизировать пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация",
                "operationId": "login",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "loginDate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/golang-pkg_internal_auth.UserForLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        },
...
        "/user/register": {
            "post": {
                "description": "Регистрация пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация",
                "operationId": "register",
                "parameters": [
                    {
                        "description": "Данные для регистрации",
                        "name": "regDate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/golang-pkg_internal_auth.UserForRegister"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/golang-pkg_internal_auth.UserForRegister"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.HackError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "golang-pkg_internal_auth.BusinessUserForRegister": {
            "type": "object",
            "required": [
                "email",
                "inn",
                "legalEntity",
                "name",
                "password",
                "phone",
                "post",
                "surname"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "inn": {
                    "type": "string"
                },
                "legalEntity": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "post": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "golang-pkg_internal_auth.UserForLogin": {
            "type": "object",
            "required": [
                "login",
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "golang-pkg_internal_auth.UserForRegister": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password",
                "phone",
                "surname"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
...
        "internal.HackError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "err": {},
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}